Create README.md
Background: 
A client from the media space approached us with a concern about a newly created Twitter account that was gaining hundreds of thousands of followers. The client suspected that the account was artificially trending obscure cryptocurrency hashtags to make them appear popular and attract genuine crypto investors. The client believed that the account was tweeting with a hashtag of a cryptocurrency, and there were a few scores to a few hundred accounts that replied or retweeted these tweets. The client suspected that all these accounts were operated by bots with 0 to few followers and that they were created just to keep the main tweeter account's tweets trending.

Problem: 
To get the required data, we needed access to Twitter API with enhanced privileges, as the data to be generated was substantial. The client offered to buy the enterprise version of the API so that we could extract all the data we wanted. However, we negotiated with Twitter and got the free version with enhanced privileges, which allowed us to extract 2 million data points as opposed to 75k data points in the basic version.

Solution: 
With access to the API, we started building a model to validate or reject the client's assumption. We explored various ways to extract the data and parsed it to extract the tweeter handles that interacted with the main account in the previous 10 days. Once we had a list of all the unique Twitter accounts, we went to each account's profile and extracted the date of creation and the number of followers.

Outcome: 
We found over 10,000 unique Twitter accounts that had interacted with the main account in the previous 10 days. When we profiled these 10,000 accounts, we found that many were created in the past 8 to 12 months and had a limited number of followers, which indicated that this might be a scam. However, the presence of over 10,000 accounts indicated that this might not be the case. We submitted a detailed analysis report with data to the client, and he agreed that this might not be a scam as initially suspected. This project was the first time we worked on uncovering potential internet fraud, and it was deeply satisfying. The client was impressed with the detailed output and awarded us with a follow-up project.